core/pyproject.toml
[project]
name = "minecraft-schematic-gen"
version = "0.1.0"
description = "Agentic interface for generating Minecraft schematics"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.3",
    "uvicorn[standard]>=0.38.0",
    "google-genai>=1.52.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
